// src/model/activity.tsx
export type MyActivityModal = {
  activityGuid: string;
  category: string;
  woNo: string;
  activityId: string;
  locationCode: string;
  locationDesc: string;
  startDatetime: string;
  endDatetime: string;
  isDeleted: string;
  createdBy: string;
  createdTimestamp: string;
  updatedBy: string;
  updatedTimestamp: string;
};